refactor: Prepare to allow Alias set on CLI
Add new parameter "alias" for search_source, which
allows us to add a new means for getting ALIAS.
This will keep the search_source func from being
tied to the ALIAS constant.

Move the instatiation of _args to __init__ this
will make it easier to use the args as state for
the app.