refactor: change names for objects in _realreq
This name change is to make things easier on the eyes. The code in
realreq should already be assumed to be private, because the `_realreq`
library is private and internal. By making all the code in `_realreq`
private, it sent mixed signals that some of it `might` be relied upon
as a stable interface. This is not the case.